add learning paths
major feature. rework on student dash. created and added LearningPaths that contain Challenges. local Question schema changed to mongodb Challenge schema. User schema updated with paths and progress. TO-DO: check to see if it wouldnt be better to avoid assigning paths to user and just appending them to the user object before we send it to the UI